Big changes to the Image Analysis library API to support animations in the App. Under the hood, now shrinks large images to speed up chart detection, face detection and image recoloring (1/35 memory for large images). Also sped up the Kuiper analysis ~160x through a different underlying data representation (that cuts App memory use by ~20MB, essentially in half!) and an optimized algorithm (no floating point calculations). The new API is much easier to integrate with Objective-C code. Renamed pipeline.h/.cpp to kokko_image.h/.cpp to reflect that this is the "gateway" header file for the image analysis.
Split the KokkoInterface.h/.mm files into two parts: the wrapper around the C++ code is now in KokkoImageWrapper.h/.mm. This is a very thin layer around the SharedImageResources and KokkoImage classes. It only exports a subset of the interface; just what the App uses. Essentially, it translates data structures between the Objective-C/UIKit and C++/openCV worlds: NSString to std::string, UIImage to cv::Mat, etc. The "logic" for the app's use of the imaging library is still in KokkoInterface, but this is now a pure Objective-C file (not Objective-C++) and deals only with the iOS side of the interface.